escrowbot-ui-main/swiftpay_api/requirements.txt
# python -m pip freeze > requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.4.1
attrs==21.2.0
Babel==2.6.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.9.3
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.11.8
cffi==1.14.6
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.8
decorator==4.3.0
defusedxml==0.7.1
Deprecated==1.2.12
distlib==0.3.4
Django==3.2.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docutils==0.14
ebaysdk==2.1.5
feedparser==5.2.1
filelock==3.4.0
Flask==1.1.2
freezegun==0.3.11
gevent==1.5.0
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
GPUtil==1.4.0
greenlet==0.4.15
html2text==2018.1.9
httplib2==0.19.1
idna==2.10
importlib-metadata==4.10.0
ipregistry==3.0.0
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==2.10.1
libsass==0.17.0
lxml==4.6.3
Mako==1.0.7
MarkupSafe==1.1.0
mysql-connector-python==8.0.26
num2words==0.5.6
numpy==1.21.0
oauthlib==3.1.1
ofxparse==0.19
opencv-python==4.5.3.56
packaging==21.0
pandas==1.1.5
passlib==1.7.1
Pillow==6.1.0
platformdirs==2.4.0
polib==1.1.0
protobuf==3.17.3
psutil==5.6.6
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydot==1.4.1
PyJWT==2.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
pyserial==3.4
python-dateutil==2.7.3
python-dotenv==0.18.0
python-stdnum==1.8.1
pytz==2019.1
pyusb==1.0.2
pywin32==301
qrcode==6.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
redis==3.5.3
reportlab==3.5.13
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rq==1.9.0
rsa==4.7.2
schedule==1.1.0
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.2
tabulate==0.8.9
typing-extensions==4.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.10.0
vobject==0.9.6.1
vonage==2.5.5
Werkzeug==0.16.1
wrapt==1.12.1
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
zeep==3.2.0
zipp==3.6.0
# python -m pip freeze > requirements.txt